# 8315789: Minor HexFormat performance improvements