This PR seeks to improve formatting of hex digits using `java.util.HexFormat` somewhat.

This is achieved getting rid of a couple of lookup tables, caching the result of `HexFormat.of().withUpperCase()`, and removing tiny allocation that happens in the `formatHex(A, byte)` method. Improvements range from 20-40% on throughput, and some operations allocate less:

```
Name                               Cnt   Base   Error   Test   Error   Unit   Diff%
HexFormatBench.appenderLower        15  1,330 ± 0,021  1,065 ± 0,067  us/op   19,9% (p = 0,000*)
  :gc.alloc.rate                    15 11,481 ± 0,185  0,007 ± 0,000 MB/sec  -99,9% (p = 0,000*)
  :gc.alloc.rate.norm               15 16,009 ± 0,000  0,007 ± 0,000   B/op -100,0% (p = 0,000*)
  :gc.count                         15  3,000          0,000         counts
  :gc.time                           3  2,000                            ms
HexFormatBench.appenderLowerCached  15  1,317 ± 0,013  1,065 ± 0,054  us/op   19,1% (p = 0,000*)
  :gc.alloc.rate                    15 11,590 ± 0,111  0,007 ± 0,000 MB/sec  -99,9% (p = 0,000*)
  :gc.alloc.rate.norm               15 16,009 ± 0,000  0,007 ± 0,000   B/op -100,0% (p = 0,000*)
  :gc.count                         15  3,000          0,000         counts
  :gc.time                           3  2,000                            ms
HexFormatBench.appenderUpper        15  1,330 ± 0,022  1,065 ± 0,036  us/op   19,9% (p = 0,000*)
  :gc.alloc.rate                    15 34,416 ± 0,559  0,007 ± 0,000 MB/sec -100,0% (p = 0,000*)
  :gc.alloc.rate.norm               15 48,009 ± 0,000  0,007 ± 0,000   B/op -100,0% (p = 0,000*)
  :gc.count                         15  0,000          0,000         counts
HexFormatBench.appenderUpperCached  15  1,353 ± 0,009  1,033 ± 0,014  us/op   23,6% (p = 0,000*)
  :gc.alloc.rate                    15 11,284 ± 0,074  0,007 ± 0,000 MB/sec  -99,9% (p = 0,000*)
  :gc.alloc.rate.norm               15 16,009 ± 0,000  0,007 ± 0,000   B/op -100,0% (p = 0,000*)
  :gc.count                         15  3,000          0,000         counts
  :gc.time                           3  2,000                            ms
HexFormatBench.toHexLower           15  0,198 ± 0,001  0,119 ± 0,008  us/op   40,1% (p = 0,000*)
  :gc.alloc.rate                    15  0,007 ± 0,000  0,007 ± 0,000 MB/sec   -0,0% (p = 0,816 )
  :gc.alloc.rate.norm               15  0,001 ± 0,000  0,001 ± 0,000   B/op  -40,1% (p = 0,000*)
  :gc.count                         15  0,000          0,000         counts
HexFormatBench.toHexLowerCached     15  0,201 ± 0,002  0,114 ± 0,001  us/op   43,0% (p = 0,000*)
  :gc.alloc.rate                    15  0,007 ± 0,000  0,007 ± 0,000 MB/sec   -0,2% (p = 0,116 )
  :gc.alloc.rate.norm               15  0,001 ± 0,000  0,001 ± 0,000   B/op  -43,1% (p = 0,000*)
  :gc.count                         15  0,000          0,000         counts
HexFormatBench.toHexUpper           15  0,146 ± 0,002  0,114 ± 0,001  us/op   21,6% (p = 0,000*)
  :gc.alloc.rate                    15  0,007 ± 0,000  0,007 ± 0,000 MB/sec    0,0% (p = 0,668 )
  :gc.alloc.rate.norm               15  0,001 ± 0,000  0,001 ± 0,000   B/op  -21,5% (p = 0,000*)
  :gc.count                         15  0,000          0,000         counts
HexFormatBench.toHexUpperCached     15  0,199 ± 0,002  0,116 ± 0,003  us/op   41,7% (p = 0,000*)
  :gc.alloc.rate                    15  0,007 ± 0,000  0,007 ± 0,000 MB/sec    0,0% (p = 0,684 )
  :gc.alloc.rate.norm               15  0,001 ± 0,000  0,001 ± 0,000   B/op  -41,7% (p = 0,000*)
  :gc.count                         15  0,000          0,000         counts
  * = significant

Invariant parameters used by above microbenchmarks:
size:   512
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must be properly reviewed (1 review required, with at least 1 [Reviewer](https://openjdk.org/bylaws#reviewer))
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue

### Issue
 * [JDK-8315789](https://bugs.openjdk.org/browse/JDK-8315789): Minor HexFormat performance improvements (**Enhancement** - P4)


### Reviewers
 * [Roger Riggs](https://openjdk.org/census#rriggs) (@RogerRiggs - **Reviewer**)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk.git pull/15591/head:pull/15591` \
`$ git checkout pull/15591`

Update a local copy of the PR: \
`$ git checkout pull/15591` \
`$ git pull https://git.openjdk.org/jdk.git pull/15591/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 15591`

View PR using the GUI difftool: \
`$ git pr show -t 15591`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk/pull/15591.diff">https://git.openjdk.org/jdk/pull/15591.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk/pull/15591#issuecomment-1708398245)