@wenshao 2023-09-08
Should we test the performance comparison of toHexDigits(long)?

I've done some work before, and I didn't find a way to perform better than a lookup table without using a lookup table.