@rgiulietti 2023-09-08
I'm not sure that micro-benchmarks are very indicative on whether a lookup table performs better than short and straightforward code.
The reason is that, once in the CPU caches, a lookup table in micro-benchmarks stays there, whereas in more realistic situations, where access is more spread out in time, it is often evicted to make room for other, more lively data.
A micro-benchmark using a lookup table shows good results because of a high rate of cache hits, whereas in other real-world workloads a lookup table might result in many cache misses on access.