@cl4es 2023-09-08
I also tried variants of this for `toHexDigits(short|int|long)` but it always ends up worse:
```
        ByteArrayLittleEndian.setChar(rep, 2, (char)(toHighHexDigit(value) << 8 | toLowHexDigit(value)));
        value >>= 8;
        ByteArrayLittleEndian.setChar(rep, 0, (char)(toHighHexDigit(value) << 8 | toLowHexDigit(value)));
```
and: 
```
        ByteArrayLittleEndian.setInt(rep, 0, toHighHexDigit(value >> 8) << 24 | toLowHexDigit(value >> 8) << 16 | toHighHexDigit(value) << 8 | toLowHexDigit(value)));
```
Only a win on `toHexDigits(byte)`:

```
Name                            Cnt  Base   Error   Test   Error  Unit   Diff%
HexFormatBench.toHexDigitsByte   15 1,992 ± 0,008  1,908 ± 0,053 us/op    4,2% (p = 0,000*)
HexFormatBench.toHexDigitsInt    15 2,476 ± 0,018  2,896 ± 0,023 us/op  -16,9% (p = 0,000*)
HexFormatBench.toHexDigitsLong   15 3,992 ± 0,052  4,229 ± 0,036 us/op   -5,9% (p = 0,000*)
HexFormatBench.toHexDigitsShort  15 2,183 ± 0,018  2,800 ± 0,162 us/op  -28,3% (p = 0,000*)
  * = significant
```
This is indicative that any win here comes from tickling the JIT the right way. I'll leave the code unchanged but add these microbenchmarks if anyone wants to attempt other improvements.